+++ python/lsst/ts/m2gui/controltab/tab_alarm_warn.py
@@ -298,7 +298,7 @@ class TabAlarmWarn(TabDefault):
     async def _callback_enable_open_loop_max_limit(self):
         """Callback of the enable button to enable the maximum limit of
         open-loop control."""
-        await run_command(self.model.enable_open_loop_max_limit)
+        await run_command(self.model.enable_open_loop_max_limit, True)
 
     def _set_signal_error(self, signal_error):
         """Set the error signal with callback functions.